+++ src/relatedProperties/propertyTypes/PropertyTypeTree.php
@@ -6,7 +6,9 @@
  * @date 30.04.2015
  */
 namespace skeeks\cms\relatedProperties\propertyTypes;
+use skeeks\cms\backend\widgets\SelectModelDialogTreeWidget;
 use skeeks\cms\components\Cms;
+use skeeks\cms\models\CmsTree;
 use skeeks\cms\relatedProperties\PropertyType;
 use yii\helpers\ArrayHelper;
 use yii\widgets\ActiveForm;
@@ -21,7 +23,21 @@ class PropertyTypeTree extends PropertyType
     public $name = "Привязка к разделу";
 
     public $is_multiple = false;
+    public $root_tree_id = null;
 
+
+    const FIELD_ELEMENT_DEFAULT                 = "selectDefault";
+    const FIELD_ELEMENT_SELECT_DIALOG           = "selectDialog";
+
+    public $fieldElement            = self::FIELD_ELEMENT_DEFAULT;
+
+    static public function fieldElements()
+    {
+        return [
+            self::FIELD_ELEMENT_DEFAULT          => \Yii::t('skeeks/cms', 'Standard selection element'),
+            self::FIELD_ELEMENT_SELECT_DIALOG    => \Yii::t('skeeks/cms','Selection in the dialog box'),
+        ];
+    }
 
     /**
      * Файл с формой настроек, по умолчанию лежит в той же папке где и компонент.
@@ -31,6 +47,10 @@ class PropertyTypeTree extends PropertyType
     public function renderConfigForm(ActiveForm $activeForm)
     {
         echo $activeForm->field($this, 'is_multiple')->checkbox(\Yii::$app->formatter->booleanFormat);
+        echo $activeForm->fieldSelect($this, 'fieldElement', static::fieldElements());
+        echo $activeForm->field($this, 'root_tree_id')->widget(
+            \skeeks\cms\widgets\formInputs\selectTree\SelectTreeInputWidget::class
+        );
     }
 
     public function attributeLabels()
@@ -38,6 +58,8 @@ class PropertyTypeTree extends PropertyType
         return array_merge(parent::attributeLabels(),
         [
             'is_multiple'  => \Yii::t('skeeks/cms','Multiple choice'),
+            'fieldElement'  => \Yii::t('skeeks/cms','Form element type'),
+            'root_tree_id'  => \Yii::t('skeeks/cms','Root partition'),
         ]);
     }
 
@@ -46,6 +68,9 @@ class PropertyTypeTree extends PropertyType
         return ArrayHelper::merge(parent::rules(),
         [
             ['is_multiple', 'boolean'],
+            ['fieldElement', 'in', 'range' => array_keys(static::fieldElements())],
+            ['fieldElement', 'string'],
+            ['root_tree_id', 'integer'],
         ]);
     }
 
@@ -66,14 +91,39 @@ class PropertyTypeTree extends PropertyType
     {
         $field = parent::renderForActiveForm();
 
-        $field->widget(
-            \skeeks\cms\widgets\formInputs\selectTree\SelectTree::className(),
-            [
-                "mode" => $this->isMultiple ? \skeeks\cms\widgets\formInputs\selectTree\SelectTree::MOD_MULTI : \skeeks\cms\widgets\formInputs\selectTree\SelectTree::MOD_SINGLE,
-                "attributeSingle" => $this->property->code,
-                "attributeMulti" => $this->property->code
-            ]
-        );
+
+        if ($this->fieldElement == static::FIELD_ELEMENT_SELECT_DIALOG) {
+
+            $options = [
+                "multiple" => $this->isMultiple ? true : false,
+            ];
+
+            if ($this->root_tree_id) {
+                $options['dialogRoute'] = ['/cms/admin-tree', 'root_id' => $this->root_tree_id];
+            }
+
+            $field->widget(
+                SelectModelDialogTreeWidget::className(),
+                $options
+            );
+        } else {
+            $rootTreeModels = [];
+            if ($this->root_tree_id) {
+                $rootTreeModels = CmsTree::findAll($this->root_tree_id);
+            }
+            $field->widget(
+                \skeeks\cms\widgets\formInputs\selectTree\SelectTreeInputWidget::className(),
+                [
+                    "multiple" => $this->isMultiple ? true : false,
+                    'treeWidgetOptions' =>
+                    [
+                        'models' => $rootTreeModels
+                    ]
+                ]
+            );
+        }
+
+
 
         return $field;
     }